refactor: move gemini files into correct folders
