+++ frontend/src/api/gemini/geminiText.ts
