+++ frontend/src/api/gemini/lib/types.ts
